--- src/lessonTwo/Office.java
@@ -1,16 +1,0 @@
-package lessonTwo;
-
-import lessonTwo.common.AbstractConnector;
-
-public class Office {
-
-    private String name;
-
-    public Office(String name) {
-        this.name = name;
-    }
-
-    public void connectToAnotherOffice(AbstractConnector connector) {
-        connector.connect(); // use abstract method
-    }
-}
