--- src/lessonTwo/common/Connector.java
@@ -1,15 +1,0 @@
-package lessonTwo.common;
-
-public class Connector {
-
-    public String createConnection(String host) {
-        var newConnection = openConnection();
-        // todo: do some work
-        return newConnection;
-    }
-
-    private String openConnection() {
-        // todo: do some work
-        return "ip address";
-    }
-}
