--- src/lessonTwo/composition/Connector.java
@@ -1,12 +1,0 @@
-package lessonTwo.composition;
-
-public class Connector {
-
-    String host;
-    Meta meta;
-
-    public Connector(String host, Meta meta) {
-        this.host = host;
-        this.meta = meta;
-    }
-}
