--- src/lessonTwo/composition/Meta.java
@@ -1,11 +1,0 @@
-package lessonTwo.composition;
-
-import java.util.HashMap;
-
-public class Meta {
-    HashMap<String, String> meta;
-
-    public Meta(HashMap<String, String> meta) {
-        this.meta = meta;
-    }
-}
