--- src/lessonTwo/solid/Office.java
@@ -1,15 +1,0 @@
-package lessonTwo.solid;
-
-import java.time.Instant;
-
-public class Office {
-
-    String name;
-    String nickname;
-    String uuid;
-    int places;
-    Instant openFrom;
-    Instant openTo;
-
-    // getters, setters, constructors
-}
