--- src/lessonTwo/solid/is/Connector.java
@@ -1,6 +1,0 @@
-package lessonTwo.solid.is;
-
-public interface Connector {
-    void connect();
-    void disconnect();
-}
