--- src/lessonTwo/solid/is/HttpConnector.java
@@ -1,7 +1,0 @@
-package lessonTwo.solid.is;
-
-public interface HttpConnector extends Connector {
-
-    void setAddress(String host);
-    void setPost(int port);
-}
