--- src/lessonTwo/solid/is/TelephoneConnector.java
@@ -1,5 +1,0 @@
-package lessonTwo.solid.is;
-
-public interface TelephoneConnector extends Connector {
-    void setNumber(String number);
-}
