--- src/lessonTwo/solid/lp/Connector.java
@@ -1,6 +1,0 @@
-package lessonTwo.solid.lp;
-
-public interface Connector {
-    void connect();
-    void disconnect();
-}
