--- src/lessonTwo/solid/lp/InternetConnector.java
@@ -1,12 +1,0 @@
-package lessonTwo.solid.lp;
-
-public class InternetConnector implements Connector {
-
-    // Constructors, getters + setters
-
-    public void connect() {}
-    public void disconnect() {}
-
-    public InternetConnector() {}
-
-}
