--- src/lessonTwo/solid/lp/MobileConnector.java
@@ -1,12 +1,0 @@
-package lessonTwo.solid.lp;
-
-public class MobileConnector implements Connector {
-
-    // Constructors, getters + setters
-
-    public void connect() {}
-    public void disconnect() {}
-
-    public MobileConnector() {}
-
-}
