--- src/lessonTwo/solid/lp/Router.java
@@ -1,14 +1,0 @@
-package lessonTwo.solid.lp;
-
-public class Router {
-
-    Connector connector;
-
-    public Router(Connector connector) {
-        this.connector = connector;
-    }
-
-    public void setConnector(Connector connector) {
-        this.connector = connector;
-    }
-}
