+++ src/lesson_1/Day.java
@@ -1,0 +1,11 @@
+package lesson_1;
+
+public enum Day {
+    MONDAY,
+    TUESDAY,
+    WEDNESDAY,
+    THURSDAY,
+    FRIDAY,
+    SATURDAY,
+    SUNDAY
+}
