+++ src/lesson_1/eq/readme.md
