+++ src/lesson_1/recordExample/UserPost.java
@@ -1,0 +1,7 @@
+package lesson_1.recordExample;
+
+public record UserPost(
+        String user,
+        String postRid,
+        int size
+) {}
