+++ src/lesson_2/common/Connector.java
@@ -1,0 +1,15 @@
+package lesson_2.common;
+
+public class Connector {
+
+    public String createConnection(String host) {
+        var newConnection = openConnection();
+        // todo: do some work
+        return newConnection;
+    }
+
+    private String openConnection() {
+        // todo: do some work
+        return "ip address";
+    }
+}
