+++ src/lesson_2/composition/Connector.java
@@ -1,0 +1,12 @@
+package lesson_2.composition;
+
+public class Connector {
+
+    String host;
+    Meta meta;
+
+    public Connector(String host, Meta meta) {
+        this.host = host;
+        this.meta = meta;
+    }
+}
