+++ src/lesson_2/composition/Meta.java
@@ -1,0 +1,11 @@
+package lesson_2.composition;
+
+import java.util.HashMap;
+
+public class Meta {
+    HashMap<String, String> meta;
+
+    public Meta(HashMap<String, String> meta) {
+        this.meta = meta;
+    }
+}
