+++ src/lesson_2/solid/Office.java
@@ -1,0 +1,15 @@
+package lesson_2.solid;
+
+import java.time.Instant;
+
+public class Office {
+
+    String name;
+    String nickname;
+    String uuid;
+    int places;
+    Instant openFrom;
+    Instant openTo;
+
+    // getters, setters, constructors
+}
