+++ src/lesson_2/solid/is/Connector.java
@@ -1,0 +1,6 @@
+package lesson_2.solid.is;
+
+public interface Connector {
+    void connect();
+    void disconnect();
+}
