+++ src/lesson_2/solid/is/HttpConnector.java
@@ -1,0 +1,7 @@
+package lesson_2.solid.is;
+
+public interface HttpConnector extends Connector {
+
+    void setAddress(String host);
+    void setPost(int port);
+}
