+++ src/lesson_2/solid/is/TelephoneConnector.java
@@ -1,0 +1,5 @@
+package lesson_2.solid.is;
+
+public interface TelephoneConnector extends Connector {
+    void setNumber(String number);
+}
