+++ src/lesson_2/solid/lp/Connector.java
@@ -1,0 +1,6 @@
+package lesson_2.solid.lp;
+
+public interface Connector {
+    void connect();
+    void disconnect();
+}
