+++ src/lesson_2/solid/lp/InternetConnector.java
@@ -1,0 +1,12 @@
+package lesson_2.solid.lp;
+
+public class InternetConnector implements Connector {
+
+    // Constructors, getters + setters
+
+    public void connect() {}
+    public void disconnect() {}
+
+    public InternetConnector() {}
+
+}
