+++ src/lesson_2/solid/lp/Router.java
@@ -1,0 +1,14 @@
+package lesson_2.solid.lp;
+
+public class Router {
+
+    Connector connector;
+
+    public Router(Connector connector) {
+        this.connector = connector;
+    }
+
+    public void setConnector(Connector connector) {
+        this.connector = connector;
+    }
+}
