--- src/n1_lesson/Day.java
@@ -1,11 +1,0 @@
-package n1_lesson;
-
-public enum Day {
-    MONDAY,
-    TUESDAY,
-    WEDNESDAY,
-    THURSDAY,
-    FRIDAY,
-    SATURDAY,
-    SUNDAY
-}
