--- src/n1_lesson/recordExample/UserPost.java
@@ -1,7 +1,0 @@
-package n1_lesson.recordExample;
-
-public record UserPost(
-        String user,
-        String postRid,
-        int size
-) {}
